+++ src/ext/java/org/opentripplanner/ext/siri/SiriTimetableSnapshotSource.java
@@ -1,6 +1,21 @@
 package org.opentripplanner.ext.siri;
 
+import static org.opentripplanner.ext.siri.TimetableHelper.createModifiedStopTimes;
+import static org.opentripplanner.ext.siri.TimetableHelper.createModifiedStops;
+import static org.opentripplanner.ext.siri.TimetableHelper.createUpdatedTripTimes;
+import static org.opentripplanner.model.PickDrop.NONE;
+import static org.opentripplanner.model.PickDrop.SCHEDULED;
+
 import com.google.common.base.Preconditions;
+import java.time.ZonedDateTime;
+import java.util.ArrayList;
+import java.util.BitSet;
+import java.util.Date;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+import java.util.TimeZone;
+import java.util.concurrent.locks.ReentrantLock;
 import org.opentripplanner.common.model.T2;
 import org.opentripplanner.model.Agency;
 import org.opentripplanner.model.FeedScopedId;
@@ -787,7 +802,7 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
         for (TripTimes tripTimes : times) {
             Trip trip = tripTimes.getTrip();
             for (TripPattern pattern : patterns) {
-                if (tripTimes.getNumStops() == pattern.getStopPattern().getStops().length) {
+                if (tripTimes.getNumStops() == pattern.numberOfStops()) {
                     if (!tripTimes.isCanceled()) {
                         /*
                           UPDATED and MODIFIED tripTimes should be handled the same way to always allow latest realtime-update
@@ -796,9 +811,9 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
 
                         cancelScheduledTrip(feedId, trip.getId().getId(), serviceDate);
 
-                        // Check whether trip id has been used for previously ADDED/MODIFIED trip message and cancel
+                        // Check whether trip id has been used for previously ADDED/MODIFIED trip message and remove
                         // previously created trip
-                        cancelPreviouslyAddedTrip(feedId, trip.getId().getId(), serviceDate);
+                        removePreviousRealtimeUpdate(feedId, trip.getId().getId(), serviceDate);
 
                         // Calculate modified stop-pattern
                         Timetable currentTimetable = getCurrentTimetable(pattern, serviceDate);
@@ -935,31 +950,28 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
     }
 
     /**
-     * Cancel previously added trip from buffer if there is a previously added trip with given trip
+     * Removes previous trip-update from buffer if there is an update with given trip
      * id (without agency id) on service date
      *
      * @param feedId feed id the trip id belongs to
      * @param tripId trip id without agency id
      * @param serviceDate service date
-     * @return true if a previously added trip was cancelled
+     * @return true if a previously added trip was removed
      */
-    private boolean cancelPreviouslyAddedTrip(final String feedId, final String tripId, final ServiceDate serviceDate) {
+    private boolean removePreviousRealtimeUpdate(final String feedId, final String tripId, final ServiceDate serviceDate) {
         boolean success = false;
 
-        final TripPattern pattern = buffer.getLastAddedTripPattern(new FeedScopedId(feedId, tripId), serviceDate);
+        FeedScopedId feedScopedTripId = new FeedScopedId(feedId, tripId);
+        final TripPattern pattern = buffer.getLastAddedTripPattern(feedScopedTripId, serviceDate);
         if (pattern != null) {
-            // Cancel trip times for this trip in this pattern
-            final Timetable timetable = buffer.resolve(pattern, serviceDate);
-            final int tripIndex = timetable.getTripIndex(tripId);
-            if (tripIndex == -1) {
-                LOG.warn("Could not cancel previously added trip {}", tripId);
-            } else {
-                final TripTimes newTripTimes = new TripTimes(timetable.getTripTimes(tripIndex));
-                newTripTimes.cancelTrip();
-                buffer.update(pattern, newTripTimes, serviceDate);
-//                buffer.removeLastAddedTripPattern(feedId, tripId, serviceDate);
-                success = true;
-            }
+            /*
+              Remove the previous realtime-added TripPattern from buffer.
+              Only one version of the realtime-update should exist
+             */
+            buffer.removeLastAddedTripPattern(feedScopedTripId, serviceDate);
+            buffer.removeRealtimeUpdatedTripTimes(pattern, feedScopedTripId, serviceDate);
+            success = true;
+
         }
 
         return success;
@@ -1018,7 +1030,7 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
             }
 
             var firstStop = tripPattern.getStop(0);
-            var lastStop = tripPattern.getStop(tripPattern.getStops().size() - 1);
+            var lastStop = tripPattern.lastStop();
 
             String siriOriginRef = monitoredVehicleJourney.getOriginRef().getValue();
 
@@ -1133,8 +1145,8 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
         }
 
 
-        var firstStop = tripPattern.getStop(0);
-        var lastStop = tripPattern.getStop(tripPattern.getStops().size() - 1);
+        var firstStop = tripPattern.firstStop();
+        var lastStop = tripPattern.lastStop();
 
         if (serviceDates.contains(journeyDate)) {
             boolean firstStopIsMatch = firstStop.getId().getId().equals(journeyFirstStopId);
@@ -1259,9 +1271,9 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
 
                 TripPattern pattern = routingService.getPatternForTrip().get(trip);
 
-                if (stopNumber < pattern.getStopPattern().getStops().length) {
+                if (stopNumber < pattern.numberOfStops()) {
                     boolean firstReportedStopIsFound = false;
-                    var stop = pattern.getStopPattern().getStops()[stopNumber - 1];
+                    var stop = pattern.getStop(stopNumber - 1);
                     if (firstStopId.equals(stop.getId().getId())) {
                         firstReportedStopIsFound = true;
                     }