+++ src/main/java/org/opentripplanner/inspector/WheelchairEdgeRenderer.java
@@ -3,6 +3,7 @@ package org.opentripplanner.inspector;
 import org.opentripplanner.inspector.EdgeVertexTileRenderer.EdgeVertexRenderer;
 import org.opentripplanner.model.WheelChairBoarding;
 import org.opentripplanner.routing.edgetype.StreetEdge;
+import org.opentripplanner.routing.edgetype.ElevatorHopEdge;
 import org.opentripplanner.routing.graph.Edge;
 import org.opentripplanner.routing.graph.Vertex;
 import org.opentripplanner.routing.vertextype.TransitStopVertex;
@@ -38,6 +39,15 @@ public class WheelchairEdgeRenderer implements EdgeVertexRenderer {
             } else {
                 attrs.color = slopePalette.getColor(pse.getMaxSlope());
                 attrs.label = String.format("%.02f", pse.getMaxSlope());
+            }
+        } else if (e instanceof ElevatorHopEdge) {
+            ElevatorHopEdge ehe = (ElevatorHopEdge) e;
+            if (!ehe.wheelchairAccessible) {
+                attrs.color = NO_WHEELCHAIR_COLOR;
+                attrs.label = "wheelchair=no";
+            } else {
+                attrs.color = Color.GREEN;
+                attrs.label = "elevator";
             }
         } else {
             return false;