+++ src/main/java/org/opentripplanner/netex/loader/parser/ResourceFrameParser.java
@@ -2,10 +2,12 @@ package org.opentripplanner.netex.loader.parser;
 
 import org.opentripplanner.netex.index.NetexEntityIndex;
 import org.rutebanken.netex.model.Authority;
+import org.rutebanken.netex.model.Branding;
 import org.rutebanken.netex.model.Operator;
 import org.rutebanken.netex.model.Organisation_VersionStructure;
 import org.rutebanken.netex.model.OrganisationsInFrame_RelStructure;
 import org.rutebanken.netex.model.ResourceFrame_VersionFrameStructure;
+import org.rutebanken.netex.model.TypesOfValueInFrame_RelStructure;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -14,14 +16,17 @@ import java.util.ArrayList;
 import java.util.Collection;
 
 class ResourceFrameParser extends NetexParser<ResourceFrame_VersionFrameStructure> {
+
     private static final Logger LOG = LoggerFactory.getLogger(ResourceFrameParser.class);
 
     private final Collection<Authority> authorities = new ArrayList<>();
     private final Collection<Operator> operators = new ArrayList<>();
+    private final Collection<Branding> brandings = new ArrayList<>();
 
     @Override
     void parse(ResourceFrame_VersionFrameStructure frame) {
         parseOrganization(frame.getOrganisations());
+        parseTypesOfValue(frame.getTypesOfValue());
 
         // Keep list sorted alphabetically
         warnOnMissingMapping(LOG, frame.getBlacklists());
@@ -33,7 +38,6 @@ class ResourceFrameParser extends NetexParser<ResourceFrame_VersionFrameStructur
         warnOnMissingMapping(LOG, frame.getOperationalContexts());
         warnOnMissingMapping(LOG, frame.getResponsibilitySets());
         warnOnMissingMapping(LOG, frame.getSchematicMaps());
-        warnOnMissingMapping(LOG, frame.getTypesOfValue());
         warnOnMissingMapping(LOG, frame.getVehicles());
         warnOnMissingMapping(LOG, frame.getVehicleEquipmentProfiles());
         warnOnMissingMapping(LOG, frame.getVehicleModels());
@@ -44,9 +48,11 @@ class ResourceFrameParser extends NetexParser<ResourceFrame_VersionFrameStructur
         verifyCommonUnusedPropertiesIsNotSet(LOG, frame);
     }
 
-    @Override void setResultOnIndex(NetexEntityIndex netexIndex) {
+    @Override
+    void setResultOnIndex(NetexEntityIndex netexIndex) {
         netexIndex.authoritiesById.addAll(authorities);
         netexIndex.operatorsById.addAll(operators);
+        netexIndex.brandingById.addAll(brandings);
     }
 
 
@@ -61,10 +67,32 @@ class ResourceFrameParser extends NetexParser<ResourceFrame_VersionFrameStructur
     private void parseOrganization(Organisation_VersionStructure element) {
         if (element instanceof Authority) {
             authorities.add((Authority) element);
-        } else if (element instanceof Operator) {
-                operators.add((Operator) element);
-        } else {
+        }
+        else if (element instanceof Operator) {
+            operators.add((Operator) element);
+        }
+        else {
             warnOnMissingMapping(LOG, element);
         }
+    }
+
+    private void parseTypesOfValue(TypesOfValueInFrame_RelStructure element) {
+
+        if (element == null || element.getValueSetOrTypeOfValue() == null) {
+            return;
+        }
+
+        for (JAXBElement<?> e : element.getValueSetOrTypeOfValue()) {
+            if (e.getValue() instanceof Branding) {
+                parseBranding((Branding) e.getValue());
+            }
+            else {
+                warnOnMissingMapping(LOG, e);
+            }
+        }
+    }
+
+    private void parseBranding(Branding branding) {
+        brandings.add(branding);
     }
 }