+++ src/main/java/org/opentripplanner/netex/mapping/StopTransferPriorityMapper.java
@@ -4,7 +4,7 @@ import javax.annotation.Nullable;
 import org.opentripplanner.model.StopTransferPriority;
 import org.rutebanken.netex.model.InterchangeWeightingEnumeration;
 
-class TransferPriorityMapper {
+class StopTransferPriorityMapper {
 
     @Nullable
     static StopTransferPriority mapToDomain(InterchangeWeightingEnumeration value) {